Add ability to mark settings as important to fail on unknown settings
Skipping unknown settings is not always the required behaviour, and most
of the time you want the opposite - fail on unknown setting.

Image that you pass some new setting, i.e. max_insert_threads, and you
assume that the INSERT will be parallel, however due to you server was
too old the setting is silently ignored.

Refs: #142